Split prompt composer view model support
