+++ scripts/file-size-hygiene-baseline.txt
@@ -6,7 +6,7 @@
 # New non-exempt tracked files above threshold_bytes fail the guard.
 # Exemptions: vendor/generated/worktree dirs and shared/core-terminal.
 134030 app/src/main/java/com/pocketshell/app/composer/PromptComposerSheet.kt
-207337 app/src/main/java/com/pocketshell/app/composer/PromptComposerViewModel.kt
+203585 app/src/main/java/com/pocketshell/app/composer/PromptComposerViewModel.kt
 139372 app/src/main/java/com/pocketshell/app/projects/FolderListScreen.kt
 180197 app/src/main/java/com/pocketshell/app/projects/FolderListViewModel.kt
 226939 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionScreen.kt
